requirements.txt
torch>=1.9.0
pandas>=1.3.0
numpy>=1.19.0
scikit-learn>=0.24.0
pyyaml>=5.4.1
tqdm>=4.61.0
matplotlib>=3.4.0
seaborn>=0.11.0